DBT-725: Fix upsert query to handle on schema changes. (#123)
As part of these changes, Make sure the upsert query works in case of column changes. I also modified the upsert query to do the following.
If a unique key is present, we will update the existing records, and when not matched, we will insert the new values.
We will always insert the new records if no unique key is present.